everything but crc cards and diagrams(9-5-26)

+++ requirements_app/src/main/java/requirements_app/domain/Project.java
@@ -7,7 +7,7 @@ public class Project {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String name;
     private String description;
@@ -29,8 +29,8 @@ public class Project {
     public Project() {
     }
 
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
